Permisije, dupuna baze

+++ BioskopAplikacija/src/servlets/KartaServlet.java
@@ -1,6 +1,7 @@
 package servlets;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -55,12 +56,24 @@ public class KartaServlet extends HttpServlet {
 			int id = Integer.parseInt(request.getParameter("id"));
 			
 			Karta karta = KartaDao.getOne(id);
-			System.out.println(karta.toString());
+		
 			Map<String, Object> data = new LinkedHashMap<>();
 			data.put("karta", karta);
 			request.setAttribute("data", data);
 			request.getRequestDispatcher("./SuccessServlet").forward(request, response);
 			
+		}
+		
+		
+		if(action.equals("getKarteZaKorisnika")) {
+			String korIme = request.getParameter("korIme");
+			System.out.println("DAJ KARTE ZA " + korIme);
+			ArrayList<Karta> karte = KartaDao.getKarteZaKorisnika(korIme);
+			
+			Map<String, Object> data = new LinkedHashMap<>();
+			data.put("karte", karte);
+			request.setAttribute("data", data);
+			request.getRequestDispatcher("./SuccessServlet").forward(request, response);
 		}
 	}
 
@@ -69,34 +82,40 @@ public class KartaServlet extends HttpServlet {
 	 */
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		String action = request.getParameter("action");
+		String ulogovaniKorisnikKorIme = (String) request.getSession().getAttribute("UlogovaniKorisnik");
+		String ulogovaniKorisnikUloga = (String) request.getSession().getAttribute("UlogovaniKorisnikUloga");
 
 		if(action.equals("delete")) {
-			int id = Integer.parseInt(request.getParameter("id"));
-			
-			if(KartaDao.delete(id)) {
-				Karta obrisanaKarta = KartaDao.getOne(id);
-				int idSedista = KartaDao.getSedisteIdZaKartu(id);
-				SedisteDao.oslobodiSediste(idSedista);
+			if(ulogovaniKorisnikUloga.equals("ADMIN")) {
+				int id = Integer.parseInt(request.getParameter("id"));
 				
-				request.getRequestDispatcher("./SuccessServlet").forward(request, response);
-				System.out.println("USPEO");
-			}else {
-				request.getRequestDispatcher("./FailureServlet").forward(request, response);
-				System.out.println("NIJE");
+				if(KartaDao.delete(id)) {
+					Karta obrisanaKarta = KartaDao.getOne(id);
+					int idSedista = KartaDao.getSedisteIdZaKartu(id);
+					SedisteDao.oslobodiSediste(idSedista);
+					
+					request.getRequestDispatcher("./SuccessServlet").forward(request, response);
+					System.out.println("USPEO");
+				}else {
+					request.getRequestDispatcher("./FailureServlet").forward(request, response);
+					System.out.println("NIJE");
+				}
 			}
 		}
 		
 		if(action.equals("add")) {
-			int idProjekcije = Integer.parseInt(request.getParameter("idProjekcije"));
-			String kupacKarteKorIme = request.getParameter("kupacKarte");
-			int sedisteId = Integer.parseInt(request.getParameter("sediste"));
-			
-			Projekcija projekcija = ProjekcijaDao.getOne(idProjekcije);
-			Sediste sediste = SedisteDao.getOne(sedisteId);
-			Korisnik kupacKarte = KorisnikDao.getOne(kupacKarteKorIme);
-			
-			if(KartaDao.add(new Karta(0, projekcija, sediste, null, kupacKarte, false))) {
-				SedisteDao.zauzmiSediste(idProjekcije, sedisteId);
+			if(!ulogovaniKorisnikKorIme.equals(null)) {
+				int idProjekcije = Integer.parseInt(request.getParameter("idProjekcije"));
+				String kupacKarteKorIme = request.getParameter("kupacKarte");
+				int sedisteId = Integer.parseInt(request.getParameter("sediste"));
+				
+				Projekcija projekcija = ProjekcijaDao.getOne(idProjekcije);
+				Sediste sediste = SedisteDao.getOne(sedisteId);
+				Korisnik kupacKarte = KorisnikDao.getOne(kupacKarteKorIme);
+				
+				if(KartaDao.add(new Karta(0, projekcija, sediste, null, kupacKarte, false))) {
+					SedisteDao.zauzmiSediste(idProjekcije, sedisteId);
+				}
 			}
 		}
