+++ BioskopAplikacija/src/servlets/KorisnikServlet.java
@@ -39,6 +39,8 @@ public class KorisnikServlet extends HttpServlet {
 	 */
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		String ulogovaniKorisnikKorIme = (String) request.getSession().getAttribute("UlogovaniKorisnik");
+		String ulogovaniKorisnikUloga = (String) request.getSession().getAttribute("UlogovaniKorisnikUloga");
+		
 
 		if(ulogovaniKorisnikKorIme == null) {
 			request.getRequestDispatcher("./OdjavaServlet").forward(request, response);
@@ -65,39 +67,51 @@ public class KorisnikServlet extends HttpServlet {
 			
 						}
 					if(action.equals("getOne")) {
+						
 						String korIme = request.getParameter("korIme");
-						Korisnik korisnik = KorisnikDao.getOne(korIme);
-						if(korisnik != null) {
+						
+						
+						if(ulogovaniKorisnikUloga.equals("ADMIN") || ulogovaniKorisnikKorIme.equals(korIme)) {
 							
-							java.util.Date datumRegistracije =  korisnik.getDatumRegistracije();
-							SimpleDateFormat sdf = new SimpleDateFormat("dd/MM/yyy HH:mm");
-							String datumReg = sdf.format(datumRegistracije);
-							
-							data.put("datumRegistracije", datumReg);
-							data.put("korisnik", korisnik);
-							request.setAttribute("data", data);
-							request.getRequestDispatcher("./SuccessServlet").forward(request, response);
+							Korisnik korisnik = KorisnikDao.getOne(korIme);
+							if(korisnik != null) {
+								
+								java.util.Date datumRegistracije =  korisnik.getDatumRegistracije();
+								SimpleDateFormat sdf = new SimpleDateFormat("dd/MM/yyy HH:mm");
+								String datumReg = sdf.format(datumRegistracije);
+								
+								data.put("datumRegistracije", datumReg);
+								data.put("korisnik", korisnik);
+								request.setAttribute("data", data);
+								request.getRequestDispatcher("./SuccessServlet").forward(request, response);
+						}
+						
 						}
 						else {
+							
 							request.getRequestDispatcher("./FailureServlet").forward(request, response);
+							
 						}
 						
 						
 					}
 					if(action.equals("getAll")) {
-						List<Korisnik> korisnici = KorisnikDao.getAll();
-					
-						ArrayList<Korisnik> neobrisaniKorisnici = new ArrayList<Korisnik>();
-						
-						for (Korisnik korisnik : korisnici) {
-							if(!korisnik.isObrisan()) {
-								neobrisaniKorisnici.add(korisnik);
-								
+						if(ulogovaniKorisnikUloga.equals("ADMIN")) {
+							List<Korisnik> korisnici = KorisnikDao.getAll();
+							
+							ArrayList<Korisnik> neobrisaniKorisnici = new ArrayList<Korisnik>();
+							
+							for (Korisnik korisnik : korisnici) {
+								if(!korisnik.isObrisan()) {
+									neobrisaniKorisnici.add(korisnik);
+									
+								}
 							}
+							data.put("korisnici", neobrisaniKorisnici);
+							request.setAttribute("data", data);
+							request.getRequestDispatcher("./SuccessServlet").forward(request, response);
 						}
-						data.put("korisnici", neobrisaniKorisnici);
-						request.setAttribute("data", data);
-						request.getRequestDispatcher("./SuccessServlet").forward(request, response);
+						
 					}
 					
 					
@@ -151,6 +165,8 @@ public class KorisnikServlet extends HttpServlet {
 	 * @see HttpServlet#doPost(HttpServletRequest request, HttpServletResponse response)
 	 */
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		String ulogovaniKorisnikKorIme = (String) request.getSession().getAttribute("UlogovaniKorisnik");
+		String ulogovaniKorisnikUloga = (String) request.getSession().getAttribute("UlogovaniKorisnikUloga");
 		String action = request.getParameter("action");
 		
 		if(action.equals("edit")) {
@@ -159,27 +175,30 @@ public class KorisnikServlet extends HttpServlet {
 			String admin = request.getParameter("admin");
 			String uloga;
 			
-			Korisnik izmenjenKorisnik;
+			if(ulogovaniKorisnikUloga.equals("ADMIN") || ulogovaniKorisnikKorIme.equals(korisnickoIme)) {
 			
-			if(admin.equals("true")) {
-				uloga = "ADMIN";
-			}
-			else {
-				uloga = "KORISNIK";
-			}
-			System.out.println("ULOGA JE : " + uloga);
-			
-			if(lozinka.equals("")) {
-				String staraLozinka = KorisnikDao.getOne(korisnickoIme).getLozinka();
-				izmenjenKorisnik = new Korisnik(korisnickoIme, staraLozinka, null, uloga, false);
+				Korisnik izmenjenKorisnik;
 				
-			}
-			else {
-				izmenjenKorisnik = new Korisnik(korisnickoIme, lozinka, null, uloga, false);
-			}
-			
-			if(KorisnikDao.update(izmenjenKorisnik)) {
-				request.getRequestDispatcher("./SuccessServlet").forward(request, response);
+				if(admin.equals("true")) {
+					uloga = "ADMIN";
+				}
+				else {
+					uloga = "KORISNIK";
+				}
+				System.out.println("ULOGA JE : " + uloga);
+				
+				if(lozinka.equals("")) {
+					String staraLozinka = KorisnikDao.getOne(korisnickoIme).getLozinka();
+					izmenjenKorisnik = new Korisnik(korisnickoIme, staraLozinka, null, uloga, false);
+					
+				}
+				else {
+					izmenjenKorisnik = new Korisnik(korisnickoIme, lozinka, null, uloga, false);
+				}
+				
+				if(KorisnikDao.update(izmenjenKorisnik)) {
+					request.getRequestDispatcher("./SuccessServlet").forward(request, response);
+				}
 			}
 			else {
 				request.getRequestDispatcher("./FailureServlet").forward(request, response);
@@ -190,8 +209,10 @@ public class KorisnikServlet extends HttpServlet {
 		
 		if(action.equals("delete")) {
 			String korisnickoIme = request.getParameter("korisnickoIme");
-			if(KorisnikDao.delete(korisnickoIme)) {
-				request.getRequestDispatcher("./SuccessServlet").forward(request, response);
+			if(ulogovaniKorisnikUloga.equals("ADMIN") && !ulogovaniKorisnikKorIme.equals(korisnickoIme)) {
+				if(KorisnikDao.delete(korisnickoIme)) {
+					request.getRequestDispatcher("./SuccessServlet").forward(request, response);
+				}
 			}
 			else {
 				request.getRequestDispatcher("./FailureServlet").forward(request, response);
